Toggle field-centric driving with start+a
Hold down the Start button on gamepad #1 and field-centric driving is controlled by the A button. If it's pressed then you get field-centric, otherwise you get robot-centric. So just tapping Start pops you over to robot-centric. Hold Start and hold A, then release Start first and you'll be back to field-centric.

This is so that if the IMU drifts too much, or if we screw up passing the heading from auto to teleop you can just tap start and be into robot-centric driving quickly.