Pass heading from auto mode to teleop mode for field-centric driving
All auto modes now automatically pass the last robot heading to the next teleop mode. This allows us to do field-centric driving after auto. 

Also added a method called setStartingHeading() to the start of every auto mode. This takes the heading of the robot, in degrees, at the start of the auto mode. This allows us to place the robot, say, facing right along the wall and still have field-centric teleop work afterwards. We must update the starting heading for each auto mode. 

Updated the field-centric code to use all of this info, and tested with my own bot. Works like a champ.