Integrated our webcam example code into auto modes
All auto modes now have access to one of two webcams. You select which one in your constructor, as shown in FTC7760FathomAutoBlueLeft. You then have access to the position of the TSE as seen when the auto mode starts in the variable `tseStartingPosition`. We can use that to set the arm position automatically later.

I added access to two webcams because I think the team will want that one day ;)

All I really did here was to package the example code we had into a class and use that during our auto opmode setup.